move engine util to engine
