+++ pgdrive/component/base_class/randomizable.py
@@ -1,4 +1,4 @@
-from pgdrive.utils.random import get_np_random
+from pgdrive.utils import get_np_random
 
 
 class Randomizable:
