+++ pgdrive/component/map/base_map.py
@@ -10,7 +10,7 @@ from pgdrive.component.blocks.pg_block import PGBlock
 from pgdrive.component.road.road import Road
 from pgdrive.component.road.road_network import RoadNetwork
 from pgdrive.utils import Config, import_pygame
-from pgdrive.utils.engine_utils import get_engine
+from pgdrive.engine.engine_utils import get_engine
 
 pygame = import_pygame()
