+++ pgdrive/engine/core/image_buffer.py
@@ -27,7 +27,7 @@ class ImageBuffer:
         parent_node: NodePath,
         frame_buffer_property=None,
     ):
-        from pgdrive.utils.engine_utils import get_engine
+        from pgdrive.engine.engine_utils import get_engine
         self.engine = get_engine()
         try:
             assert self.engine.win is not None, "{} cannot be made without use_render or use_image".format(
