+++ pgdrive/envs/pgdrive_env_v2_minimal.py
@@ -7,7 +7,7 @@ from pgdrive.envs.pgdrive_env_v2 import PGDriveEnvV2
 from pgdrive.obs.observation_base import ObservationBase
 from pgdrive.obs.state_obs import LidarStateObservation
 from pgdrive.utils import Config
-from pgdrive.utils.engine_utils import get_engine
+from pgdrive.engine.engine_utils import get_engine
 from pgdrive.utils.math_utils import norm, clip
 
 DISTANCE = 50
