+++ pgdrive/manager/base_manager.py
@@ -8,7 +8,7 @@ class BaseManager(Randomizable):
     Managers should be created and registered after launching BaseEngine
     """
     def __init__(self):
-        from pgdrive.utils.engine_utils import get_engine, engine_initialized
+        from pgdrive.engine.engine_utils import get_engine, engine_initialized
         assert engine_initialized(), "You should not create manager before the initialization of BaseEngine"
         self.engine = get_engine()
         Randomizable.__init__(self, self.engine.global_random_seed)
