+++ pgdrive/manager/spawn_manager.py
@@ -10,7 +10,7 @@ from pgdrive.component.lane.straight_lane import StraightLane
 from pgdrive.constants import CollisionGroup
 from pgdrive.utils import Config, get_np_random
 from pgdrive.utils.coordinates_shift import panda_position, panda_heading
-from pgdrive.utils.engine_utils import get_engine
+from pgdrive.engine.engine_utils import get_engine
 from pgdrive.utils.scene_utils import rect_region_detection
