+++ pgdrive/manager/traffic_manager.py
@@ -9,7 +9,7 @@ from pgdrive.component.road.road import Road
 from pgdrive.constants import TARGET_VEHICLES, TRAFFIC_VEHICLES, OBJECT_TO_AGENT, AGENT_TO_OBJECT
 from pgdrive.manager.base_manager import BaseManager
 from pgdrive.utils import norm, merge_dicts
-from pgdrive.utils.engine_utils import get_engine
+from pgdrive.engine.engine_utils import get_engine
 
 BlockVehicles = namedtuple("block_vehicles", "trigger_road vehicles")
