+++ pgdrive/tests/test_env/test_build_city.py
@@ -3,7 +3,7 @@ import pygame
 from pgdrive import PGDriveEnv
 from pgdrive.component.map.city_map import CityMap
 from pgdrive.obs.top_down_renderer import draw_top_down_map
-from pgdrive.utils.engine_utils import initialize_engine, close_engine
+from pgdrive.engine.engine_utils import initialize_engine, close_engine
 
 
 def _t(num_blocks):
