+++ pgdrive/tests/test_policy/test_idm_policy.py
@@ -6,7 +6,7 @@ from pgdrive.envs.base_env import BASE_DEFAULT_CONFIG
 from pgdrive.envs.pgdrive_env import PGDriveEnvV1_DEFAULT_CONFIG
 from pgdrive.policy.idm_policy import IDMPolicy
 from pgdrive.utils import Config
-from pgdrive.utils.engine_utils import initialize_engine
+from pgdrive.engine.engine_utils import initialize_engine
 
 
 def _create_vehicle():
