+++ pgdrive/utils/__init__.py
@@ -1,8 +1,7 @@
 from pgdrive.utils.config import Config, merge_config_with_unknown_keys, merge_config
 from pgdrive.utils.coordinates_shift import panda_heading, panda_position, pgdrive_heading, pgdrive_position
 from pgdrive.utils.cutils import import_cutils
-from pgdrive.utils.engine_utils import get_engine, get_global_config
 from pgdrive.utils.math_utils import safe_clip, clip, norm, distance_greater, safe_clip_for_small_array, Vector
 from pgdrive.utils.utils import is_mac, import_pygame, recursive_equal, setup_logger, merge_dicts, \
     concat_step_infos, is_win
-from pgdrive.utils.random import get_np_random
+from pgdrive.utils.random import get_np_random, random_string
